Add stein to RDO Dashboards and remove Ocata
Although pipelines are still WIP, i think we can enable Stein in the
dashboards already.

Also, i think it's a good time to remove Ocata as the pipeline is not
longer running and it's just in Extended Maintenance.

Finally, this patch is also adding the dashboard for DLRN promotions for
rocky, which was not added in the past and stein.

Change-Id: Ib2f84029307817def88fb2199df434a969205fdf