Move rdo-dashboards to CentOS Stream 9
This patch implements some changes required to run rdo-dashboards in
CS9:

- Moves from dashing to smashing which is the "spiritual successor to
  dashing" [1]
- Fix packages installation commands.
- Fix virtualenv option

Note that, the command to start it changes from dashing to smashing so
some other changes in scripts or systemd definition may be needed.

I'm also adding a functional job to test the rdo-dashboards deployment
following similar approach to the production ansible role [2] (with some
adjustments).

[1] https://github.com/Smashing/smashing
[2] https://github.com/rdo-infra/ansible-role-rdo_dashboards

Change-Id: Ia99f8650759aa8be438721d2792732c115a65a24